backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
numpy==1.26.4
scipy==1.13.0
scikit-learn==1.4.2
opencv-python-headless==4.9.0.80
Pillow==10.3.0
python-multipart==0.0.9
asyncio==3.4.3
